requirements.txt
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
SQLAlchemy>=2.0.36
APScheduler==3.10.4
vosk==0.3.45
requests==2.31.0
python-dateutil==2.8.2
pydub==0.25.1
pytz==2024.1

